modified:   MapData.csv 	modified:   Map_Service/db/migrate/20150309042521_add_attributes_to_mapspaces.rb 	modified:   Map_Service/db/schema.rb

+++ Map_Service/db/migrate/20150309042521_add_attributes_to_mapspaces.rb
@@ -1,8 +1,8 @@
 class AddAttributesToMapspaces < ActiveRecord::Migration
   def change
       add_column :mapspaces, :name, :string
-      add_column :mapspaces, :latitude, :integer
-      add_column :mapspaces, :longitude, :integer
+      add_column :mapspaces, :latitude, :double
+      add_column :mapspaces, :longitude, :double
       add_column :mapspaces, :north_adjoin, :integer
       add_column :mapspaces, :south_adjoin, :integer
       add_column :mapspaces, :east_adjoin, :integer
